fix: type mismatch due to inconsistent type resolution
It's possible to get a type mismatch when our dependencies use
different versions of upstream crates, even if the types are
structurally the same. This occurred for example in two places:

* passing a `Denomination` to `to_string_with_denomination`
* calling `to_bech32` on an `XOnlyPublicKey`

causing a failure to resolve the correct types as fn arguments.

To work around this, we get the correct type by first converting to
another common rust type, u64 in the case of `Amount`, or String in the
case of `XOnlyPublicKey`. It's not clear whether this is a better
approach in the long run than simply pinning dependencies as was
done previously.